:sparkles: feat : 딜러와 참가자의 게임 결과(수익률) 가져오는 기능 구현

+++ src/main/java/domain/player/Participant.java
@@ -32,7 +32,7 @@ public class Participant extends Player {
     }
     
     @Override
-    public double calculateProfit(int betAmount) {
+    public double calculateProfit(double betAmount) {
         return getState().calculateProfit(betAmount);
     }
 }
