+++ src/main/java/domain/state/BlackJack.java
@@ -8,7 +8,7 @@ public class BlackJack extends Finished {
     }
     
     @Override
-    public double calculateProfit(int betAmount) {
+    public double calculateProfit(double betAmount) {
         return betAmount * 1.5;
     }
 }
