+++ src/main/java/domain/state/Running.java
@@ -8,7 +8,7 @@ public abstract class Running extends State {
     }
     
     @Override
-    public double calculateProfit(int betAmount) {
+    public double calculateProfit(double betAmount) {
         throw new IllegalStateException("아직 배팅 금액을 계산할 수 없는 상태입니다.");
     }
