+++ src/main/java/domain/state/Stay.java
@@ -8,7 +8,7 @@ public class Stay extends Finished {
     }
     
     @Override
-    public double calculateProfit(int betAmount) {
+    public double calculateProfit(double betAmount) {
         return betAmount;
     }
 }
